+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v1.0.1 (2026-08-31)
+### Featured Changes
+* Added legacy option for AqualinkD to use Temp sensors for PPM,PH,ORP vs the using Lux sensors.
+* Fixed some light & conversion issues
+
+## v1.0.0 (2026-08-30)
+### Featured Changes
+* Fist release, supports both AqualinkD and AquachemD
 
 ## v0.0.1 (2026-07-3)
 ### Featured Changes